Relocate advice specific fixtures
