+++ ui_tests/caseworker/conftest.py
@@ -7,7 +7,7 @@ from selenium.webdriver.support import expected_conditions
 from selenium.webdriver.support.wait import WebDriverWait
 
 from core.constants import CaseStatusEnum
-from ui_tests.caseworker.pages.advice import FinalAdvicePage, RecommendationsAndDecisionPage, TeamAdvicePage
+from ui_tests.caseworker.pages.advice import FinalAdvicePage, TeamAdvicePage
 from ui_tests.caseworker.pages.case_page import CasePage, CaseTabs
 from ui_tests.caseworker.pages.goods_queries_pages import StandardGoodsReviewPages, OpenGoodsReviewPages
 from ui_tests.caseworker.pages.teams_pages import TeamsPages
@@ -32,7 +32,6 @@ from ui_tests.caseworker.fixtures.add_a_picklist import (  # noqa
     add_a_standard_advice_picklist,
     add_a_report_summary_picklist,
 )
-from ui_tests.caseworker.pages.advice import UserAdvicePage
 from ui_tests.caseworker.pages.generate_decision_documents_page import GeneratedDecisionDocuments
 from ui_tests.caseworker.pages.generate_document_page import GeneratedDocument
 from ui_tests.caseworker.pages.give_advice_pages import GiveAdvicePages
@@ -179,9 +178,6 @@ def prepare_for_final_review(driver, api_test_client):  # noqa
     )
 
 
-@when("I click move case forward")
-@when("I click submit recommendation")
-@when("I click save and publish to exporter")
 @when("I click save")
 @when("I click preview")
 @when("I click confirm")
@@ -406,17 +402,6 @@ def select_template_by_name(driver, template_name):  # noqa
     GeneratedDocument(driver).select_document_template_by_name(template_name)
 
 
-@then("I see the licence number on the SIEL licence preview")
-def should_see_licence_number_on_siel_licence_preview(driver, context):  # noqa
-    text = GeneratedDocument(driver).get_document_preview_text()
-    assert f"{context.reference_code}-01" in text
-
-
-@then(parsers.parse('I see that "{item_name}" is "{value}" on the SIEL licence preview'))  # noqa
-def should_see_item_on_siel_licence_preview(driver, item_name, value):  # noqa
-    assert value == GeneratedDocument(driver).get_item_from_siel_document_preview(item_name)
-
-
 @when("I go to the documents tab")  # noqa
 def click_documents(driver):  # noqa
     CasePage(driver).change_tab(CaseTabs.DOCUMENTS)
@@ -476,16 +461,6 @@ def approve_application_objects(context, api_test_client, decision):  # noqa
 
     api_test_client.cases.create_user_advice(context.case_id, data)
     api_test_client.cases.create_team_advice(context.case_id, data)
-
-
-@when("I go to the final advice page by url")  # noqa
-def final_advice_page(driver, context, internal_url):  # noqa
-    driver.get(
-        internal_url.rstrip("/")
-        + "/queues/00000000-0000-0000-0000-000000000001/cases/"
-        + context.case_id
-        + "/final-advice/"
-    )
 
 
 @when("I click edit flags link")  # noqa
@@ -560,139 +535,11 @@ def i_create_an_standard_advice_picklist(context, add_a_standard_advice_picklist
     context.standard_advice_query_picklist_question_text = add_a_standard_advice_picklist["text"]
 
 
-@when("I click the recommendations and decision tab")
-def click_on_recommendations_and_decision_tab(driver, context):  # noqa
-    CasePage(driver).change_tab("advice")
-
-
-@when("I click make recommendation")
-def click_make_recommendation_button(driver):  # noqa
-    RecommendationsAndDecisionPage(driver).click_make_recommendation()
-
-
-@when("I click approve all")
-def click_approve_all(driver):  # noqa
-    RecommendationsAndDecisionPage(driver).click_approve_all()
-
-
-@when("I click refuse all")
-def click_refuse_all(driver):  # noqa
-    RecommendationsAndDecisionPage(driver).click_refuse_all()
-
-
-@when(parsers.parse('I select refusal criteria "{criteria}"'))
-def select_refusal_criteria(driver, criteria):  # noqa
-    page = RecommendationsAndDecisionPage(driver)
-
-    for crit in criteria.split(","):
-        page.select_refusal_criteria(crit.strip())
-
-
-@when(parsers.parse('I select countries "{countries}"'))
-def select_countries(driver, countries):  # noqa
-    page = RecommendationsAndDecisionPage(driver)
-
-    for country in countries.split(","):
-        page.select_country(country.strip())
-
-
-@when(parsers.parse('I enter "{reasons}" as the overall reason'))
-@when(parsers.parse('I enter "{reasons}" as the reasons for approving'))
-def enter_reasons_for_approving(driver, reasons, context):  # noqa
-    RecommendationsAndDecisionPage(driver).enter_reasons_for_approving(reasons)
-
-
-@when(parsers.parse('I enter "{reasons}" as the reasons for refusal'))
-def enter_reasons_for_refusal(driver, reasons, context):  # noqa
-    RecommendationsAndDecisionPage(driver).enter_reasons_for_refusal(reasons)
-
-
-@when(parsers.parse('I enter "{licence_condition}" as the licence condition'))
-def enter_licence_condition(driver, licence_condition, context):  # noqa
-    RecommendationsAndDecisionPage(driver).enter_licence_condition(licence_condition)
-
-
-@when(parsers.parse('I enter "{instructions}" as the instructions for the exporter'))
-def enter_instructions_for_exporter(driver, instructions, context):  # noqa
-    RecommendationsAndDecisionPage(driver).enter_instructions_for_exporter(instructions)
-
-
-@when(parsers.parse('I enter "{footnote}" as the reporting footnote'))
-def enter_reporting_footnote(driver, footnote, context):  # noqa
-    RecommendationsAndDecisionPage(driver).enter_reporting_footnote(footnote)
-
-
-@then(parsers.parse('I see "{reasons}" as the overall reason'))
-@then(parsers.parse('I see "{reasons}" as the reasons for approving'))
-def should_see_reasons_for_approving(driver, reasons, context):  # noqa
-    assert RecommendationsAndDecisionPage(driver).get_reasons_for_approving() == reasons
-
-
-@then(parsers.parse('I see "{reasons}" as the reasons for refusal'))
-def should_see_reasons_for_refusal(driver, reasons, context):  # noqa
-    assert RecommendationsAndDecisionPage(driver).get_reasons_for_refusal() == reasons
-
-
-@then(parsers.parse('I see "{criteria}" as the refusal criteria'))
-def should_see_refusal_criteria(driver, criteria):  # noqa
-    page = RecommendationsAndDecisionPage(driver)
-
-    assert all(crit == criteria for crit in page.get_refusal_criteria())
-
-
-@then(parsers.parse('I see "{licence_condition}" as the licence condition'))
-def should_see_licence_condition(driver, licence_condition, context):  # noqa
-    assert RecommendationsAndDecisionPage(driver).get_licence_condition() == licence_condition
-
-
-@then(parsers.parse('I see "{instructions}" as the instructions for the exporter'))
-def should_see_instructions_for_exporter(driver, instructions, context):  # noqa
-    assert RecommendationsAndDecisionPage(driver).get_instructions_for_exporter() == instructions
-
-
-@then(parsers.parse('I see "{footnote}" as the reporting footnote'))
-def should_see_reporting_footnote(driver, footnote, context):  # noqa
-    assert RecommendationsAndDecisionPage(driver).get_reporting_footnote() == footnote
-
-
 @when(parsers.parse('I expand the details for "{details_text}"'))
 def expand_details(driver, details_text):  # noqa
-    driver.find_element_by_xpath(f"//span[contains(text(), '{details_text}')]").click()
-
-
-@then(parsers.parse('I should see my recommendation for "{countries}" with "{reasons}"'))
-def should_see_recommendation(driver, countries, reasons):  # noqa
-    text = driver.find_element_by_xpath("//main[@class='govuk-main-wrapper']//*").text
-    assert "Approved by 2 lite-team" in text
-    for country in countries.split(","):
-        assert country.strip() in text
-    assert reasons.strip() in text
-
-
-@then(parsers.parse('I see there are no recommendations from "{team}"'))
-def should_not_see_recommendation(driver, team, context):  # noqa
-    assert f"{team} has approved" not in Shared(driver).get_text_of_body()
-
-
-@then("I am asked what my recommendation is")
-def should_ask_for_recommendation(driver):  # noqa
-    assert Shared(driver).get_text_of_heading().text == "What is your recommendation?"
-
-
-@when("I click on the user advice tab")  # noqa
-def i_click_on_view_advice(driver, context):  # noqa
-    CasePage(driver).change_tab(CaseTabs.USER_ADVICE)
-
-
-@when("I select all items in the user advice view")  # noqa
-def click_items_in_advice_view(driver, context):  # noqa
-    context.number_of_advice_items_clicked = UserAdvicePage(driver).click_on_all_checkboxes()
-
-
-@when(parsers.parse("I choose to '{option}' the licence"))  # noqa
-def choose_advice_option(driver, option, context):  # noqa
-    GiveAdvicePages(driver).click_on_advice_option(option)
-    context.advice_data = []
+    driver.find_element_by_xpath(
+        f"//details[@class='govuk-details']/summary/span[contains(text(), '{details_text}')]"
+    ).click()
 
 
 @when(parsers.parse("I import text from the '{option}' picklist"))  # noqa
@@ -712,11 +559,6 @@ def write_note_text_field(driver, text, context):  # noqa
 @when(parsers.parse("I select that a footnote is not required"))  # noqa
 def write_note_text_field(driver, text, context):  # noqa
     GiveAdvicePages(driver).select_footnote_not_required()
-
-
-@when("I combine all user advice")  # noqa
-def combine_all_advice(driver):  # noqa
-    UserAdvicePage(driver).click_combine_advice()
 
 
 @given("I create a letter paragraph picklist")  # noqa
